ENH: Suggestion to clearly split CLI click commands in __init__.py and reusable Python API in dtool_lookup_client.utils. Latter still based on dtool_lookup_gui.LookupClient, but that's exchangable without affecting the exposed API.
