+++ dtool_lookup_client/config.py
@@ -1,0 +1,10 @@
+import dtoolcore.utils
+
+CONFIG_PATH = dtoolcore.utils.DEFAULT_CONFIG_PATH
+
+DTOOL_LOOKUP_SERVER_URL_KEY = "DTOOL_LOOKUP_SERVER_URL"
+DTOOL_LOOKUP_SERVER_TOKEN_KEY = "DTOOL_LOOKUP_SERVER_TOKEN"
+DTOOL_LOOKUP_SERVER_USERNAME_KEY = "DTOOL_LOOKUP_SERVER_USERNAME"
+DTOOL_LOOKUP_SERVER_PASSWORD_KEY = "DTOOL_LOOKUP_SERVER_PASSWORD"
+
+DTOOL_LOOKUP_CLIENT_IGNORE_SSL_KEY = "DTOOL_LOOKUP_CLIENT_IGNORE_SSL"
